#34 chore: 회원가입 폼 관련 타입 코드 폴더 이동
