+++ src/constants/form/signup.ts
@@ -1,13 +1,4 @@
-export interface signupProps {
-  [index: string]: string | number
-  email: string
-  password: string
-  name: string
-  dsId: string
-  dsName: string
-}
-
-export type signupValidationProps = (obj: signupProps) => string
+import { signupProps, signupValidationProps } from '@/types'
 
 export const SIGNUP_VALIDATION: signupValidationProps = ({
   name,
