+++ src/types/form/signup.ts
@@ -1,0 +1,10 @@
+export interface signupProps {
+  [index: string]: string | number
+  email: string
+  password: string
+  name: string
+  dsId: string
+  dsName: string
+}
+
+export type signupValidationProps = (obj: signupProps) => string
